[Issue 38] turn elixir project into umbrella app
