+++ elixir/.gitignore
@@ -6,4 +6,3 @@ erl_crash.dump
 /fractals
 /images
 /logs
-/test/images
