--- elixir/.projectile
@@ -1,3 +1,0 @@
--/_build
--/deps
--/.credo.exs
