+++ elixir/apps/fractals/.formatter.exs
@@ -1,0 +1,7 @@
+[
+  inputs: [
+    "lib/**/*.{ex,exs}",
+    "test/**/*.{ex,exs}",
+    "mix.exs"
+  ]
+]
