+++ elixir/apps/fractals/config/config.exs
@@ -1,0 +1,5 @@
+use Mix.Config
+
+config :logger, backends: [:console], compile_time_purge_level: :info
+
+config :fractals, unimplemented: [:newton, :nova]
