+++ elixir/apps/fractals/lib/fractals/colorizer/black_and_white_and_gray.ex
