+++ elixir/apps/fractals/lib/fractals/colorizer_worker.ex
