+++ elixir/apps/fractals/lib/fractals/escape_time/burning_ship.ex
