+++ elixir/apps/fractals/lib/fractals/escape_time/helpers.ex
