+++ elixir/apps/fractals/lib/fractals/escape_time/mandelbrot.ex
