+++ elixir/apps/fractals/lib/fractals/escape_time_worker.ex
