+++ elixir/apps/fractals/lib/fractals/image_magick.ex
