+++ elixir/apps/fractals/lib/fractals/output_manager.ex
