+++ elixir/apps/fractals/lib/fractals/reporters/stdout.ex
