+++ elixir/apps/fractals/test/fractals/escape_time/burning_ship_test.exs
