+++ elixir/apps/fractals/test/inputs/full_params.yml
@@ -8,4 +8,5 @@ c:          3.14+4.13i
 z:          4.4+1.1i
 r:          9.9+3.3i
 p:          0.3+0.5i
+output_directory: "test/images"
 output_filename: "the-output.png"
