+++ elixir/apps/fractals/test/inputs/integration_mandelbrot.yml
