+++ elixir/apps/fractals/test/inputs/partial_params.yml
@@ -1,0 +1,3 @@
+fractal: Julia
+color:   Blue
+output_directory: "test/images"
