+++ elixir/config/config.exs
@@ -1,5 +1,3 @@
 use Mix.Config
 
-config :logger, backends: [:console], compile_time_purge_level: :info
-
-config :fractals, unimplemented: [:newton, :nova]
+import_config "../apps/*/config/config.exs"
