+++ elixir/mix.exs
@@ -1,41 +1,16 @@
-defmodule Fractals.Mixfile do
+defmodule Fractals.MixProject do
   use Mix.Project
 
   def project do
     [
-      app: :fractals,
-      version: "0.0.1",
-      elixir: "~> 1.3",
-      escript: [main_module: Fractals.CLI],
-      build_embedded: Mix.env() == :prod,
+      apps_path: "apps",
       start_permanent: Mix.env() == :prod,
       deps: deps(),
-      dialyzer: [ignore_warnings: ".dialyzer-ignore-warnings"],
-      aliases: aliases()
+      dialyzer: [ignore_warnings: ".dialyzer-ignore-warnings"]
     ]
-  end
-
-  def application do
-    [applications: [:logger, :yaml_elixir], mod: {Fractals, []}]
   end
 
   defp deps do
-    [
-      {:complex, github: "jdfrens/elixir-complex", ref: "51e2804"},
-      {:credo, "~> 0.10.0", only: [:dev, :test]},
-      {:dialyxir, "~> 1.0.0-rc.2", only: [:dev, :test], runtime: false},
-      {:earmark, "~> 1.2", override: true},
-      {:gen_stage, "~> 0.14"},
-      {:inflex, "~> 1.10.0"},
-      {:mogrify, "~> 0.6.1"},
-      {:uuid, "~> 1.1"},
-      {:yaml_elixir, "~> 2.1.0"}
-    ]
-  end
-
-  defp aliases do
-    [
-      all: ["test", "mix format --check-formatted", "credo --strict"]
-    ]
+    []
   end
 end
