--- elixir/test/inputs/partial_params.yml
@@ -1,2 +1,0 @@
-fractal: Julia
-color:   Blue
